requirements.txt
apache-airflow-providers-apache-spark
apache-airflow-providers-mongo==3.0.0
apache-airflow-providers-google
gcsfs
pymongo[srv]
pyspark
requests
pytz
python-dateutil